backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy
python-multipart
pydantic-settings
pydantic[email]
